Proper PropertyCollection caching
A PropertyCollection now keeps tracks of its children and updates them
whenever a new property is added to a class. This enables faster
querying for a particular property. Previously, this query would have
caused walking the entire PropertyCollection chain and merging several
hashes.

Internally, a PropertyCollection references its properties using strings
because these are GCed. Assigning unsafe input to a SmartProperties
enabled object could otherwise lead to excisive memory usage because
PropertyCollection#key? would have to convert to a symbol which would
not be GCed.