Use equality test for nil as a last resort
I ran into an issue when using properties that are
ActiveRecord::Relation objects - essentially what was happening was
extra queries being made to the database during smart property
validation. It turns out that ActiveRecord (and probably other things)
can have side effects when using the built-in equality methods.

This PR changes the order of the nil-checking for properties and uses
the `#nil?` method as the default now; if a `NoMethodError` is raised,
then we can safely do a nil singleton equality check since the object
being validated is most likely descending from BasicObject.