Fixing PropertieCollection to deal with prepended modules
There is an issue when prepending a module to a class that
has SmartProperties. When SmartProperties::PropertyCollection.for
is called, it will identify the class as a parent of itself. This
will cause a recursive ancestor check.

To fix this, I added a simple check in the parent identification
to ensure that the ancestor is not the base class.