Fix check for including `ModuleMethods`
The parameter passed to `Module.included` is always a `Module` so the
check for deciding if the library should inject the `ModuleMethods` was
not doing any checks at all.

Since a `Class` is always also a `Module`, the check should check to see
if the parameter is not a `Class` instead. This makes sure that
`ModuleMethods` is only included into modules and never to classes.